add router to event detail
