+++ src/app/event-detail/[TokkenAddress]/layout.tsx
@@ -1,0 +1,12 @@
+export default function Layout({
+      children,
+    }: {
+      children: React.ReactNode,
+    }) {
+      return (
+        <section>
+          <nav></nav>
+          {children}
+        </section>
+      );
+    }
