--- src/app/event-detail/[nft-id].tsx
@@ -1,6 +1,0 @@
-import { useRouter } from 'next/router'
-
-export default function Page() {
-  const router = useRouter()
-  return <p>Post: {router.query.slug}</p>
-}
